// sitat-fe-workspace/projects/app-contratti/src/app/modules/models/scheda-sicurezza/scheda-sicurezza.model.ts
export interface SchedaSicurezzaEntry {
    codGara?: number;
    codLotto?: number;
    direttoreOperativo?: string;
    numIniz?: number;
    pianoSicurezza?: string;
    tutor?: string;
}

export interface SchedaSicurezzaInsertForm {
    codGara?: number;
    codLotto?: number;
    direttoreOperativo?: string;
    numIniz?: number;
    pianoSicurezza?: string;
    tutor?: string;
}